64bit compile fix

+++ gui/asystem.cpp
@@ -490,7 +490,7 @@ void	aObject::setTexture(unsigned long newHandle )
 }
 
 
-void aObject::setColor( unsigned long newColor, bool bRecurse )
+void aObject::setColor( uint32_t newColor, bool bRecurse )
 {
 	for ( int i = 0; i < 4; i++ )
 	{
